Updates prototype for 'index_str_index', now returns 'uint8_t *'.

+++ test/test_index_sym.c
@@ -393,27 +393,27 @@ test_sym_str_index
    test_assert_critical(sym != NULL);
    
    char * t1 = "ATGCATCGTAGCA";
-   int32_t s1[] = {0,3,2,1,0,3,1,2,3,0,2,1,0};
+   uint8_t s1[] = {0,3,2,1,0,3,1,2,3,0,2,1,0};
 
    char * t2 = "ATGTCNATNNCCA";
-   int32_t s2[] = {0,3,2,3,1,4,0,3,4,4,1,1,0};
+   uint8_t s2[] = {0,3,2,3,1,4,0,3,4,4,1,1,0};
 
    char * t3 = "GTKGA0ACGPT!$";
-   int32_t s3[] = {2,3,4,2,0,4,0,1,2,4,3,4,4};
+   uint8_t s3[] = {2,3,4,2,0,4,0,1,2,4,3,4,4};
 
-   int32_t * s = sym_str_index(t1, sym);
+   uint8_t * s = sym_str_index(t1, sym);
    test_assert_critical(s != NULL);
-   test_assert(memcmp(s, s1, 13*sizeof(int32_t)) == 0);
+   test_assert(memcmp(s, s1, 13*sizeof(uint8_t)) == 0);
    free(s);
 
    s = sym_str_index(t2, sym);
    test_assert_critical(s != NULL);
-   test_assert(memcmp(s, s2, 13*sizeof(int32_t)) == 0);
+   test_assert(memcmp(s, s2, 13*sizeof(uint8_t)) == 0);
    free(s);
 
    s = sym_str_index(t3, sym);
    test_assert_critical(s != NULL);
-   test_assert(memcmp(s, s3, 13*sizeof(int32_t)) == 0);
+   test_assert(memcmp(s, s3, 13*sizeof(uint8_t)) == 0);
    free(s);
 
    sym_free(sym);
